% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc.R
\docType{data}
\name{Data_Counts}
\alias{Data_Counts}
\title{Weekly notified diseases reported by Ministry of Health, Nutrition
and Indigenous Medicine in Sri Lanka}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18772 rows and 18 columns.
}
\source{
Weekly epidemiological reports published by the Epidemiological Unit, Ministry of Health, Sri Lanka
}
\usage{
Data_Counts
}
\description{
\describe{
\item{ID}{row id}
\item{Year}{Year}
\item{Week}{Week number to locate the week within a year}
\item{StartData}{Starting date of the week}
\item{EndData}{End date of the week}
\item{Division}{Districts of Sri Lanka}
\item{Dengue}{Weekly Dengue counts}
\item{Dysentery}{Weekly Dysentery counts}
\item{Encephalitis}{Weekly Encephalitis counts}
\item{Enteric Fever}{Weekly Enteric Fever Counts}
\item{Food Poisioning}{Weekly Food Poisioning Counts}
\item{Leptospirosis}{Weekly Leptospirosis Counts}
\item{Typhus Fever}{Weekly Typhus Fever Counts}
\item{Viral Hepatistis}{Weekly Viral Hepatistis Counts}
\item{Human Rabies}{Weekly Human Rabies Counts}
\item{Chickenpox}{Weekly Chickenpox Counts}
\item{Meningitis}{Weekly Meningitis Counts}
\item{Leishmaniasis}{Weekly Leishmaniasis Counts}
}
}
\examples{
data(Data_Counts)
head(Data_Counts)
summary(Data_Counts)
}
\keyword{datasets}
