% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fixed.R
\name{plot_fixed}
\alias{plot_fixed}
\title{Fix plot}
\usage{
plot_fixed(p, size = 0.95)
}
\arguments{
\item{p}{a plot object}

\item{size}{size of the viewport}
}
\value{
draws the plot
}
\description{
Fix plot
}
\examples{
p <- ggplot2::ggplot(mtcars,
ggplot2::aes(wt, mpg)) +

ggplot2::geom_point()
plot_fixed(p)
}
