% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingExcelFiles.R
\name{Processed_Data}
\alias{Processed_Data}
\title{Single data table from generated excel files}
\usage{
Processed_Data(processed_data_path)
}
\arguments{
\item{processed_data_path}{location of generated excel files}
}
\value{
a table in which columns contain ID, Year, week, StartDate, EndDate and disease counts.
Rows represents divisions of Sri Lanka.
}
\description{
Weekly notified Dengue, Dysentery, Encephalitis, Enteric Fever,
Food Poisioning, Leptospirosis, Typhus Fever, Viral Hepatistis, Human Rabies,
Chickenpox, Meningitis and Leishmaniasis cases in 26 districts of Sri Lanka
from 2006/week-52 to 2020/week-51
}
