% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{Reading_EpidemiologicalReports}
\alias{Reading_EpidemiologicalReports}
\title{Extract data table from epidemiological reports}
\usage{
Reading_EpidemiologicalReports(
  path,
  startdate,
  year,
  week,
  Cname,
  DeletingRows,
  TotalColmns,
  PercentageColmns,
  DeleteColmns
)
}
\arguments{
\item{path}{location of stored epidemiological reports}

\item{startdate}{Start date of the week}

\item{year}{year of the start date}

\item{week}{week number to locate week within a year}

\item{Cname}{Column names of data table}

\item{DeletingRows}{Unnecessary rows of data table}

\item{TotalColmns}{Total columns of data table}

\item{PercentageColmns}{Columns corresponds to Timeliness and Completeness of epidemiological pdf report}

\item{DeleteColmns}{Columns corresponds to cumulative disease cases}
}
\value{
list of tables
A table corresponds to a week of the considered year in which each column contains
weekly counts of 11 diseases and rows represent divisions of Sri Lanka.
}
\description{
Weekly notified Dengue, Dysentery, Encephalitis, Enteric Fever,
Food Poisioning, Leptospirosis, Typhus Fever, Viral Hepatistis, Human Rabies,
Chickenpox, Meningitis and Leishmaniasis cases in 26 districts of Sri Lanka
from 2006/week-52 to 2020/week-51
}
