% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc.R
\docType{data}
\name{Data_Counts}
\alias{Data_Counts}
\title{Weekly notified diseases reported by Ministry of Health, Nutrition
and Indigenous Medicine in Sri Lanka}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19006 rows and 18 columns.
}
\source{

}
\usage{
Data_Counts
}
\description{

}
\examples{
data(Data_Counts)
head(Data_Counts)
summary(Data_Counts)
}
\keyword{datasets}
